Images added, appearance improved, enums updated, readme updated.
There are 4 new images:
*AddStudent.png
*RemoveStudent.png
*SearchStudent.png
*UpdateStudent.png

The main menu window now has 4 icons to make it more understandable. Also, the window to create a student is now almost working.

The enums have been updated to represent correctly the options of the school.

Finally, the readme have been modified to show the images banks in a new sub-header, and a new bank have been added.